Move PyAudioObservableBackgroundDeviceMonitor into audio_capture.model.monitor
